+++ hostyarg/internal/expect_tests/expect_tests.go
@@ -1,4 +1,4 @@
-package testrunner
+package expect_tests
 
 import (
 	"bufio"
