--- hostyarg/internal/testrunner/testrunner_test.go
@@ -1,1 +1,0 @@
-package testrunner
